Document text transfer milestone and split UI components
Update the public README to reflect the current Android text-transfer
milestone instead of the older ping/request-response prototype.

The README now describes the real working flow:
- Android devices discover each other with NSD
- sender creates a text offer
- receiver can accept or decline
- accepted text is transferred over local Ktor HTTP
- receiver can view and copy the received text

Also replace the README demo GIF with the latest working Sync360 demo and
update the screenshots guide so the hero demo is described as the current
text-transfer flow.

Refactor Send and Receive UI into smaller feature components:
- move Send screen status, text input, file placeholder, and nearby-device
  section into dedicated composables
- move Receive idle, incoming text offer, and received text UI into dedicated
  composables
- keep screen files focused on collecting state and wiring ViewModel actions

Clean up small networking/readability details:
- configure Ktor client timeouts at the HttpClient level
- handle request timeout exceptions alongside socket/connect timeouts
- make text offer server response handling explicit with a when branch

Add a receiver-side guard for incoming text offers so a device that is already
busy or still showing received content declines new offers instead of replacing
the current receive state.
