+++ shared/src/commonMain/kotlin/com/liftley/sync360/data/network/http/client/TextOfferException.kt
@@ -1,0 +1,3 @@
+package com.liftley.sync360.data.network.http.client
+
+class TextOfferException(response: String) : Exception("Offer status: $response")
