--- shared/src/commonMain/kotlin/com/liftley/sync360/data/remote/server/serverTextResponse/OfferException.kt
@@ -1,3 +1,0 @@
-package com.liftley.sync360.data.remote.server.serverTextResponse
-
-class OfferException(response: String) : Exception("Offer status: $response")
